[LCR-208] refactor: ApiResponseType으로 type 옮기기
