+++ src/mocks/HomeMocks.ts
@@ -1,4 +1,5 @@
-import { BannerItemType, HotItemType, PopUpItemType } from '@/types/HomeScreen';
+import { GetHotPopUpItemsResponse, GetPopUpItemsResponse } from '@/types/api/ApiResponseType';
+import { BannerItemType } from '@/types/HomeScreenType';
 
 const Images = {
   popUp01: require('@/assets/images/home/popUp01.webp'),
@@ -30,7 +31,7 @@ export const bannerItems: BannerItemType[] = [
   },
 ];
 
-export const hotItems: HotItemType[] = [
+export const hotItems: GetHotPopUpItemsResponse = [
   {
     popupId: 1,
     popupName: '여자친구 Season of Memories 여자친구 Season of Memories',
@@ -65,7 +66,7 @@ export const hotItems: HotItemType[] = [
   },
 ];
 
-export const popUpItems: PopUpItemType[] = [
+export const popUpItems: GetPopUpItemsResponse = [
   {
     popupId: 1,
     popupName: '여자친구 Season of Memories 여자친구 Season of Memories',
