+++ src/types/HomeScreenType.ts
@@ -1,0 +1,5 @@
+export type BannerItemType = {
+  image: number;
+  title: string;
+  subTitle: string;
+};
